Rydder opp i Auth klassene og filene
